% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search_terms-data.R
\docType{data}
\name{advanced_search_terms}
\alias{advanced_search_terms}
\title{Clinicaltrials.gov advanced search terms}
\format{A data frame with three columns: the key to use in advanced search queries \code{keys}, a description of the search term \code{description}, and a link to the help document \code{help}.}
\source{
\url{http://clinicaltrials.gov/ct2/search/advanced}
}
\usage{
data(advanced_search_terms)
}
\description{
Clinicaltrials.gov advanced search terms. Their keys, descriptions, and links to help documents.
}
\examples{
advanced_search_terms
\dontrun{browseURL(advanced_search_terms["age", "help"])}
}
\keyword{datasets}

