% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gather_results}
\alias{gather_results}
\title{Parses results from an xml object downloaded from clinicaltrials.gov}
\usage{
gather_results(parsed)
}
\arguments{
\item{parsed}{A parsed XML object, as returned by \code{XML::xmlParse}}
}
\value{
A list of \code{data.frame}s, participant flow, baseline data,
  outcome results
}
\description{
Results of a clinical study are stored in a particular way. This reads and
organizes the information and returns it as a list of dataframes. Throws an error if the xml has no \code{clinical_results} node.
}
\keyword{Internal}

