% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStudyField.r
\name{getStudyField}
\alias{getStudyField}
\title{Returns values from selected API fields for a large set of study records}
\usage{
getStudyField(expr, fields, max_rnk = NULL)
}
\arguments{
\item{expr}{Search Expression (sep with +, use AND, OR, ...)}

\item{fields}{Study Fields (see https://clinicaltrials.gov/api/info/study_fields_list for a list of available fields)}

\item{max_rnk}{Maximum Rank (default is min(numberReturns, 20))}
}
\value{
data.frame
}
\description{
https://clinicaltrials.gov/api/gui/ref/api_urls
}
\examples{
getStudyField(expr = 'COVID+AND+hydroxychloroquine', fields = c("NCTId"))
getStudyField(expr = 'COVID+AND+hydroxychloroquine', fields = c("NCTId"),max_rnk=50)
getStudyField(expr = 'COVID+AND+hydroxychloroquine', fields = c("NCTId", "BriefTitle"))
}
\author{
JuG
}
