% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid.pull.R
\name{pmid.pull}
\alias{pmid.pull}
\title{Retrieves list of PubMed IDs associated with trial on ClinicalTrials.gov.}
\usage{
pmid.pull(nct_id)
}
\arguments{
\item{nct_id}{list of NCT IDs. Can be passed from clinicaltrials_download function}
}
\value{
A \code{data.frame} of NCT ID and the associated PMIDs
}
\description{
Retrieves list of PubMed IDs associated with trial on ClinicalTrials.gov.
}
