% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubinfo.R
\name{pubinfo}
\alias{pubinfo}
\title{Parses publication information from downloaded file from
PubMed.}
\usage{
pubinfo(p)
}
\arguments{
\item{p}{dataframe (taken from \code{pmid.pull}) or vector of PMIDs}
}
\value{
A \code{data.frame} of parsed publication information
}
\description{
Retrieves publication details given an input of PubMed IDs.
Input is typically the output from the pmid.pull function.
}
